+++ ERapi/Models/ProductModel.cs
@@ -6,8 +6,8 @@ namespace ER.Models
     public class ProductModel : BaseEntity
     {
         public string Name { get; set; }
-        public double UnitValue { get; set; }
-        public double Cost { get; set; }
+        public decimal UnitValue { get; set; }
+        public decimal Cost { get; set; }
 
     }
 }
